started to impl cvar linker, fixed event tests that did not work

+++ include/sani/core/cvar/cvar_file.hpp
@@ -21,7 +21,7 @@ namespace sani {
 		/// Returns the filename.
 		const String& getFilename() const;
 		/// Returns the count of lines.
-		const size_t getLinesCount() const;
+		const uint32 getLinesCount() const;
 
 		/// Returns given line at given index.
 		const String& lineAtIndex(const size_t index) const;
